change version and rename mcmc script
