+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "phylox"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
   { name="Remie Janssen", email="remiejanssen92@gmail.com" },
 ]
