+++ src/phylox/generators/mcmc/__init__.py
@@ -1,1 +1,1 @@
-from .MCMCSamplerVerticalMoves import *
+from .base import *
